put logging config in its own py file
