+++ README.md
@@ -162,51 +162,23 @@ Install it the way you wish.  These were my notes.
 1. The software allows users to define a time window for each aggregation record. Specify the time in _msec_ with the `--window <size>` offering.. TThe default is stored in  `set.py` . The time is in milliseconds. 
 
 ## Usage Notes:
-* Linux users can set the execute bit on main.py and run the main.py directly without the `python3` part.
-    ```
-    chmod +x main.py
-    ```
+*Linux users can set the execute bit on main.py and run the main.py directly without the `python3` part.
+```
+chmod +x main.py
+```
 
-### Sample: reading from file
-Reads from a pcap and writes to dataset.csv
+## Sample: Read from pcap file
+Reads from a Razi pcap file and writes to dataset.csv
+```
+python3 main.py --sourcefile Razi_15012021.pcap
+```
 
-`python3 main.py --sourcefile Razi_15012021.pcap`
-
-    
-### Capturing pcap files with tshark
+## Capturing pcap files with tshark
 Try this
 ```
 sudo tshark  -i eth0 -a duration:120 -w /tmp/foo.pcap -F pcap
 ```
 
-# Source Code
-* The source tree is formatted with _black_ in _Visual Studio Code_ extension
-
-# performance
-This progam makes use of 5 cores, 4 for python Python and one for tshark
-
-These tests were run on a slightly slower higher core, 16 core xeon 2.2Ghz from SSD. 
-Note that their performance is about the same as the 4 core run where we merged detectors and services to have fewer processes
-
-| Sample  | sample file size | real time                      |  analyzed packets  | time windows | sample period | python | 
-| ------- | ---------------- | -------------------------------|  ----------------  | ------------ | ------------- | ------ |
-| Crylock |   143,446,091 B  | real:1:43 user:1:40 sys:0:15   | n/a                | n/a | 10.04 | tshark (only) |
-| Crylock |   143,446,091 B  | real:1:47 user:7:14 sys:2:36   | 128778 @ 1259/sec  | 121 | 10:04 | cpython | 
-| Crylock |   143,446,091 B  | real:3:07 user 11:39 sys:1:21  | 128778 @ 754/sec   | 121 | 10:04 | pypy    | 
-| Maze    |   767,491,552 B  |                                | 573523 @ 1106/sec  | 111 | 09:21 | cpython |
-
-This benchmark was for 2 queue 3 python process version **prior** to adding a back the queue between detectors and services.
-Adding that topic **degraded** performance by 10% on a quad core because it added a 5th process.
-
-| Sample  | sample file size | real time                      |  analyzed packets  | time windows | sample period | python |
-| ------- | ---------------- | -------------------------------|  ----------------  | ------------ | ------------- | ------ |
-| Crylock |   143,446,091 b  | real:1:47 user:6:07 sys:1:22   | 128778 @ 1201/sec  | 121 | 10:04 | cpython |
-| Maze    | 1,045,083,415 b  | real:11:21 user:38:38 sys:8:33 | 770,987 @ 1131/sec | 94  | 7:59  | cpython |
-
-1. Analysis times are linear with the number of packets processed
-1. Tested with ransomware samples from unavarra.es some of which may have originated on other sites.
-
-# Random notes
 ## Zombie processes
 You will end up with one zombie python3 process if you `ctrl-c` the command line you ran this under.
 Run some version of this:
@@ -215,3 +187,31 @@ Run some version of this:
 pkill -f tshark
 pkill -f python3
 ```
+# performance
+This progam makes use of 5 cores, 4 for python Python and one for tshark.  
+It maxes out the cores so hyperthreaded cores will not count towards performance.
+
+These tests were run on a slightly slower 16 core xeon 2.2Ghz from SSD. 
+
+| Sample  | sample file size | real time                      |  analyzed packets  | time windows | sample period | python | 
+| ------- | ---------------- | -------------------------------|  ----------------  | ------------ | ------------- | ------ |
+| Crylock |   143,446,091 B  | real:1:43 user:1:40 sys:0:15   | n/a                | n/a | 10.04 | tshark (only) |
+| Crylock |   143,446,091 B  | real:1:47 user:7:14 sys:2:36   | 128778 @ 1259/sec  | 121 | 10:04 | cpython | 
+| Crylock |   143,446,091 B  | real:3:07 user 11:39 sys:1:21  | 128778 @ 754/sec   | 121 | 10:04 | pypy    | 
+| Maze    |   767,491,552 B  |                                | 573523 @ 1106/sec  | 111 | 09:21 | cpython |
+
+This benchmark was for 2-queue 3-python process version.  It was a test to see how much impact the queues vs the uplift of having extra processors.  For this test we removed the queue between detectors and services.
+
+| Sample  | sample file size | real time                      |  analyzed packets  | time windows | sample period | python |
+| ------- | ---------------- | -------------------------------|  ----------------  | ------------ | ------------- | ------ |
+| Crylock |   143,446,091 b  | real:1:47 user:6:07 sys:1:22   | 128778 @ 1201/sec  | 121 | 10:04 | cpython |
+| Maze    | 1,045,083,415 b  | real:11:21 user:38:38 sys:8:33 | 770,987 @ 1131/sec | 94  | 7:59  | cpython |
+
+1. Analysis times are linear with the number of packets processed
+1. Tested with ransomware samples from unavarra.es some of which may have originated on other sites.
+1. Running the 5 process (4 queue) version on quad core machines results in  **degraded** performance by 10%.  This is because we are CPU bound and have more processes that cores.
+
+
+# Source Code 
+The source tree is formatted with _black_ in _Visual Studio Code_ extension
+
